Rename BarcodeScanner file as BarcodeScannerScreen and the BarcodeScannerScreen compose into BarcodeScanner.
