+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/barcodescanner/BarcodeScanningFragment.kt
@@ -44,7 +44,7 @@ class BarcodeScanningFragment : BaseFragment() {
         viewModel.permissionState.observe(viewLifecycleOwner) { permissionState ->
             view.setContent {
                 WooThemeWithBackground {
-                    BarcodeScanner(
+                    BarcodeScannerScreen(
                         codeScanner = codeScanner,
                         permissionState = permissionState,
                         onResult = { granted ->
